Fixed typing case on parameterized inductive types
Due to the intersection of Redex properly handling binding and
dependency, must be careful when decomposing and rearranging binding
contexts.
Previously, I was not careful, and as a result, dependent case analysis
on parameterized inductive types would (non-deterministically, perhaps
due to hygiene algorithm?) fail to type check when they should.